[Refactor] 모달 함수들 모듈화 - 모달 함수 모듈화 및 연결 - naver-api.js 코드 축소
