+++ src/utils/modal.js
@@ -1,0 +1,97 @@
+import { addColorToModalFav } from '../utils/add-color-to-modal-fav';
+
+export function createBookModal() {
+  const modal = document.createElement('dialog');
+  modal.className = 'modal';
+  modal.setAttribute('aria-modal', 'true');
+  modal.style.display = 'none';
+
+  modal.innerHTML = `
+    <article class="modalContent bookModal">
+      <h2 class="sr-only">도서 정보 안내창</h2>
+      <header class="bookHeader">
+        <button class="closeModal" aria-label="닫기">&times;</button>
+        <div class="bookImageWrapper">
+          <img src="" alt="책 표지" class="bookCover" width="117" height="137" />
+        </div>
+      </header>
+      <section class="bookDetails" data-isbn="">
+        <h3 class="sr-only">도서 상세정보</h3>
+        <p class="bookTitle">제목</p>
+        <p class="bookAuthorPublisher">
+          <span class="bookAuthor">작가</span> | <span class="bookPublisher">출판사</span>
+        </p>
+        <p class="bookPrice">가격</p>
+        <p class="bookDescription">설명 없음</p>
+        <button class="favoriteButton" type="button" aria-label="즐겨찾기 추가/삭제">
+          <svg
+            class="favoriteIcon"
+            xmlns="http://www.w3.org/2000/svg"
+            id="Layer_1"
+            data-name="Layer 1"
+            viewBox="0 0 24 24"
+            aria-hidden="true"
+            >
+            <path
+                d="M16.074,3.781c-1.39-1.106-2.703-2.152-3.691-3.333L12.006-.002l-.384,.444c-1.491,1.722-1.905,3.821-2.238,5.508-.405,2.051-.654,3.037-1.884,3.037-.198,0-.35-.069-.492-.224-.607-.663-.687-2.586-.594-3.643l.123-1.403-.98,1.011c-1.418,1.463-4.057,4.186-4.057,8.771,0,5.79,4.71,10.5,10.5,10.5s10.5-4.71,10.5-10.5c0-4.603-3.267-7.204-6.426-9.719Zm-6.902,18.048c-1.56-1.562-1.56-4.101-.022-5.639l2.851-2.522,2.828,2.502c1.559,1.561,1.559,4.1,0,5.659-1.559,1.559-4.097,1.561-5.657,0Zm6.643,.365c1.642-1.962,1.563-4.888-.301-6.752l-3.513-3.11-3.536,3.13c-1.843,1.842-1.921,4.766-.279,6.731-3.343-1.473-5.685-4.813-5.685-8.694,0-3.233,1.37-5.396,2.888-7.122,.036,1.015,.221,2.338,.882,3.06,.329,.359,.753,.549,1.229,.549,2.107,0,2.476-1.867,2.865-3.844,.294-1.492,.624-3.164,1.655-4.595,.984,1.066,2.177,2.016,3.431,3.014,2.974,2.368,6.049,4.816,6.049,8.937,0,3.882-2.343,7.222-5.686,8.695Z"
+            />
+          </svg>
+        </button>
+      </section>
+    </article>
+  `;
+  document.body.appendChild(modal);
+  return modal;
+}
+
+export function openBookModal(modal, card) {
+  modal.querySelector('.bookTitle').textContent = card.dataset.title || '제목 없음';
+  modal.querySelector('.bookAuthor').textContent = card.dataset.author || '작가 없음';
+  modal.querySelector('.bookPublisher').textContent = card.dataset.publisher || '출판사 없음';
+  modal.querySelector('.bookPrice').textContent = card.dataset.discount + '원' || '가격 없음';
+  modal.querySelector('.bookCover').src = card.dataset.image || '';
+  modal.querySelector('.bookDescription').textContent = card.dataset.description || '설명 없음';
+  modal.querySelector('.bookDetails').dataset.isbn = card.dataset.isbn || '';
+  modal.dataset.isbn = card.dataset.isbn;
+
+  modal.style.display = 'flex';
+  addColorToModalFav(modal);
+
+  const closeBtn = modal.querySelector('.closeModal');
+  if (closeBtn) closeBtn.focus();
+
+  document.body.style.overflow = 'hidden';
+
+  // 포커스 트랩 함수 호출
+  trapFocus(modal);
+}
+
+export function modalClose(modal) {
+  modal.querySelector('.bookDetails').scrollTop = 0;
+  modal.style.display = 'none';
+  document.body.style.overflow = '';
+}
+
+export function trapFocus(modal) {
+  const focusableElement = modal.querySelectorAll('button');
+  const firstElement = focusableElement[0];
+  const lastElement = focusableElement[focusableElement.length - 1];
+
+  modal.addEventListener('keydown', (e) => {
+    if (e.key === 'Tab') {
+      if (e.shiftKey) {
+        // Shift + Tab
+        if (document.activeElement === firstElement) {
+          e.preventDefault();
+          lastElement.focus();
+        }
+      } else {
+        // Tab
+        if (document.activeElement === lastElement) {
+          e.preventDefault();
+          firstElement.focus();
+        }
+      }
+    }
+  });
+}
